requirement.txt
python-dotenv
streamlit 
gtts
google.generativeai
flask
Pillow